XStream in config util
git-svn-id: https://www.objectdefinitions.com/od-repos/odrepository/trunk@1240 899a1914-839d-7b46-afe5-da559765d5e5

+++ src/main/java/od/configutil/FileSourceAndSink.java
@@ -115,12 +115,16 @@ public class FileSourceAndSink implements ConfigSink, ConfigSource {
     
                 LogMethods.log.debug("Written: " + tempConfigFile);
 
-                LogMethods.log.debug("About to delete " + backupFile);
-                backupFile.delete();
+                if ( backupFile.exists()) {
+                    LogMethods.log.debug("About to delete " + backupFile);
+                    backupFile.delete();
+                }
 
-                LogMethods.log.debug("About to rename old config: " + configFile + " to " + backupFile);
-                if (!configFile.renameTo(backupFile)) {
-                    throw new IOException("Unable to rename old config: " + configFile + " to " + backupFile);
+                if ( configFile.exists()) {
+                    LogMethods.log.debug("About to rename old config: " + configFile + " to " + backupFile);
+                    if (!configFile.renameTo(backupFile)) {
+                        throw new IOException("Unable to rename old config: " + configFile + " to " + backupFile);
+                    }
                 }
                 
                 LogMethods.log.debug("About to rename temp config: " + tempConfigFile + " to " + configFile);
